templates: add flatcar-sysext template
This template allows to use a plain Flatcar image from the Flatcar
release server without building with the image-builder.

This template will consume Kubernetes systemd-sysext image from the
flatcar/sysext-bakery release artifacts.

This template will emit a /run/reboot-required file flag when a new
Kubernetes release has been downloaded and that we need to coordinate
the reboot of the node using Kured for example.

Signed-off-by: Mathieu Tortuyaux <mtortuyaux@microsoft.com>
Co-authored-by: Kai Lüke <pothos@users.noreply.github.com>